+++ module/iron_control_api/src/freeform/mod.rs
@@ -1,4 +1,6 @@
-/// Parser for provider-key + invite-email paste blocks.
-pub mod providers_and_invites;
+/// Parser for invite email-address paste blocks.
+pub mod invites;
+/// Parser for `provider: key` paste blocks.
+pub mod providers;
 /// Parser for usage-policy paste blocks.
 pub mod usage_policy;
